Build 3.3 (#12)
* Update IncomingMessage.cs

* Implemented Missing Messages :hammer:

* Build 3.2.2

* Update WineCellar.cs

* :fire: Hotfixes

* Some basic changes

* :fire: Hotfixes! No more high CPU usage anymore! Yay!

* Update fetchCalendar.cs

* :fire: Hotfixes! Patch 6

* Create errors-build-3.2.2-5.txt

* Build 3.2.3

* :fire: Patch 1 (3.2.3)

* Small fixes + added /size command for vips and deleted pots in nexus

* Seb was a noob NVM

* Quick Update

-the crown of frog king and mad jester robe woks now
-dyno bot behav remake

* Build 3.2.4 (rev 0)

* Build 3.2.4 (rev 1)

* :fire: Hotfix: fixed gods portal not despawn issue.

* :fire: Hotfix: prodlike gods loot.

* Small fixes :)

* :fire: Hotfix: abilities now work.

* :fire: Hotfix: fixed colors on chat.

* :fire: Hotfix: improved chat color algorithm.

* Build 3.2.5 (rev 0)

* Build 3.2.5 (rev 1)

* Fixed Merchants

* :fire: Hotfix: fixed chat color issue.

* :fire: Hotfix: added support to 'double' values.

* Loot Structure Update

Highlands :white_check_mark:
Lowlands :white_check_mark:
Frog King :white_check_mark: (also updated mana drain bomb so its not raping your mana)
Midlands :white_check_mark:

* Build 3.2.6 (rev 0)

* Build 3.2.7 (WIP)

* Build 3.2.7 (rev 0)

* Build 3.2.7 (rev 1)

* Build 3.2.8 (rev 0)

* Small Fixes (Abilities fully working)

* Update 3.2.8.1 (rev 0)

* Added Glow for VIPS

* Update 3.8.2.1 (rev 1)

* Update 3.8.2.2 (rev 2)

* Delete VS reports

* [TS] Build 3.2.8.2 (rev 0)

* [TS] Build 3.2.8.2 (rev 1)

* Fixed Bars

* Build 3.2.8.3 (rev 0)

* Build 3.2.8.3 (rev 1)

* Build 3.2.9

* udl and abbys???

idk if they work or not since i cant get localhost to wok

* woops forgot to make it embedded

* Last old client changes

* Last server side changes (WIP new integration of unity client)

* Unity: WIP new client

* Build 3.2.10-12 (VPN implemented)

+++ server/common/models/FameStats.cs
@@ -194,42 +194,42 @@ namespace LoESoft.Core
             return ret.ToArray();
         }
 
-        private static Tuple<string, string, Func<FameStats, DbChar, int, bool>, Func<double, double>>[] bonusDat = new[] {
+        private static Tuple<string, string, Func<FameStats, DbChar, double, bool>, Func<double, double>>[] bonusDat = new[] {
             Tuple.Create("Ancestor", "First death of any of your characters",
-                new Func<FameStats, DbChar, int, bool>(
+                new Func<FameStats, DbChar, double, bool>(
                     (fStats, character, baseFame) =>
                 character.CharId < 2),
                 new Func<double, double>(f => f * 0.1 + 20)
             ),
 
             Tuple.Create("Pacifist", "Never shot a bullet which hit an enemy",
-                new Func<FameStats, DbChar, int, bool>(
+                new Func<FameStats, DbChar, double, bool>(
                     (fStats, character, baseFame) =>
                 fStats.ShotsThatDamage == 0),
                 new Func<double, double>(f => f * 0.25)
             ),
             //Legacy Builder???
             Tuple.Create("Thirsty", "Never drank a potion from inventory",
-                new Func<FameStats, DbChar, int, bool>(
+                new Func<FameStats, DbChar, double, bool>(
                     (fStats, character, baseFame) =>
                 fStats.PotionsDrunk == 0),
                 new Func<double, double>(f => f * 0.25)
             ),
             Tuple.Create("Mundane", "Never used special ability (requires level 20)",
-                new Func<FameStats, DbChar, int, bool>(
+                new Func<FameStats, DbChar, double, bool>(
                     (fStats, character, baseFame) =>
                 character.Level == 20 && fStats.SpecialAbilityUses == 0),
                 new Func<double, double>(f => f * 0.25)
             ),
             Tuple.Create("Boots on the Ground", "Never teleported",
-                new Func<FameStats, DbChar, int, bool>(
+                new Func<FameStats, DbChar, double, bool>(
                     (fStats, character, baseFame) =>
                 fStats.Teleports == 0),
                 new Func<double, double>(f => f * 0.25)
             ),
 
             Tuple.Create("Tunnel Rat", "Completed every dungeon type",
-                new Func<FameStats, DbChar, int, bool>(
+                new Func<FameStats, DbChar, double, bool>(
                     (fStats, character, baseFame) =>
                 fStats.PirateCavesCompleted > 0 &&
                 fStats.UndeadLairsCompleted > 0 &&
@@ -245,79 +245,79 @@ namespace LoESoft.Core
             ),
 
             Tuple.Create("Enemy of the Gods", "More than 10% of kills are gods (requires level 20)",
-                new Func<FameStats, DbChar, int, bool>(
+                new Func<FameStats, DbChar, double, bool>(
                     (fStats, character, baseFame) =>
                 character.Level == 20 && (double)fStats.GodKills / (fStats.GodKills + fStats.MonsterKills) > 0.1),
                 new Func<double, double>(f => f * 0.1)
             ),
             Tuple.Create("Slayer of the Gods", "More than 50% of kills are gods (requires level 20)",
-                new Func<FameStats, DbChar, int, bool>(
+                new Func<FameStats, DbChar, double, bool>(
                     (fStats, character, baseFame) =>
                 character.Level == 20 && (double)fStats.GodKills / (fStats.GodKills + fStats.MonsterKills) > 0.5),
                 new Func<double, double>(f => f * 0.1)
             ),
 
             Tuple.Create("Oryx Slayer", "Dealt Killing blow to Oryx",
-                new Func<FameStats, DbChar, int, bool>(
+                new Func<FameStats, DbChar, double, bool>(
                     (fStats, character, baseFame) =>
                 fStats.OryxKills > 0),
                 new Func<double, double>(f => f * 0.1)
             ),
 
             Tuple.Create("Accurate", "Accuracy of better than 25% (requires level 20)",
-                new Func<FameStats, DbChar, int, bool>(
+                new Func<FameStats, DbChar, double, bool>(
                     (fStats, character, baseFame) =>
                 character.Level == 20 && (double)fStats.ShotsThatDamage / fStats.Shots > 0.25),
                 new Func<double, double>(f => f * 0.1)
             ),
             Tuple.Create("Sharpshooter", "Accuracy of better than 50% (requires level 20)",
-                new Func<FameStats, DbChar, int, bool>(
+                new Func<FameStats, DbChar, double, bool>(
                     (fStats, character, baseFame) =>
                 character.Level == 20 && (double)fStats.ShotsThatDamage / fStats.Shots > 0.5),
                 new Func<double, double>(f => f * 0.1)
             ),
             Tuple.Create("Sniper", "Accuracy of better than 75% (requires level 20)",
-                new Func<FameStats, DbChar, int, bool>(
+                new Func<FameStats, DbChar, double, bool>(
                     (fStats, character, baseFame) =>
                 character.Level == 20 && (double)fStats.ShotsThatDamage / fStats.Shots > 0.75),
                 new Func<double, double>(f => f * 0.1)
             ),
 
             Tuple.Create("Explorer", "More than 1 million tiles uncovered",
-                new Func<FameStats, DbChar, int, bool>(
+                new Func<FameStats, DbChar, double, bool>(
                     (fStats, character, baseFame) =>
                 fStats.TilesUncovered > 1000000),
                 new Func<double, double>(f => f * 0.05)
             ),
             Tuple.Create("Cartographer", "More than 4 million tiles uncovered",
-                new Func<FameStats, DbChar, int, bool>(
+                new Func<FameStats, DbChar, double, bool>(
                     (fStats, character, baseFame) =>
                 fStats.TilesUncovered > 4000000),
                 new Func<double, double>(f => f * 0.05)
             ),
 
             Tuple.Create("Team Player", "More than 100 party member level ups",
-                new Func<FameStats, DbChar, int, bool>(
+                new Func<FameStats, DbChar, double, bool>(
                     (fStats, character, baseFame) =>
                 fStats.LevelUpAssists > 100),
                 new Func<double, double>(f => f * 0.1)
             ),
             Tuple.Create("Leader of Men", "More than 1000 party member level ups",
-                new Func<FameStats, DbChar, int, bool>(
+                new Func<FameStats, DbChar, double, bool>(
                     (fStats, character, baseFame) =>
                 fStats.LevelUpAssists > 1000),
                 new Func<double, double>(f => f * 0.1)
             ),
 
             Tuple.Create("Doer of Deeds", "More than 1000 quests completed",
-                new Func<FameStats, DbChar, int, bool>(
+                new Func<FameStats, DbChar, double, bool>(
                     (fStats, character, baseFame) =>
                 fStats.QuestsCompleted > 1000),
                 new Func<double, double>(f => f * 0.05)
             ),
 
             Tuple.Create("Friend of the Cubes", "Never killed a cube (requires level 20)",
-                new Func<FameStats, DbChar, int, bool>(
+                new Func<FameStats, DbChar, double, bool>(
                     (fStats, character, baseFame) =>
                 character.Level == 20 && fStats.CubeKills == 0),
                 new Func<double, double>(f => f * 0.05)
@@ -331,7 +331,7 @@ namespace LoESoft.Core
                     yield return Tuple.Create(i.Item1, i.Item2, i.Item4(character.Fame));
 
             //Well Equiped
-            var bonus = character.Items.Take(4).Where(x => x != -1).Sum(x => data.Items[(ushort) x].FameBonus) / 100.0;
+            var bonus = character.Items.Take(4).Where(x => x != -1).Sum(x => data.Items[(ushort)x].FameBonus) / 100.0;
             if (bonus > 0)
                 yield return Tuple.Create("Well Equipped", "Bonus for equipment", character.Fame * bonus);
 
@@ -340,7 +340,7 @@ namespace LoESoft.Core
                 yield return Tuple.Create("First Born", "Best fame of any of your previous incarnations", character.Fame * 0.1);
         }
 
-        public int CalculateTotalFame(EmbeddedData data, DbClassStats stats, DbChar chr, int baseFame, out bool firstBorn)
+        public double CalculateTotalFame(EmbeddedData data, DbClassStats stats, DbChar chr, double baseFame, out bool firstBorn)
         {
             double bonus = 0;
 
@@ -370,11 +370,11 @@ namespace LoESoft.Core
                 bonus = Math.Floor(bonus) + (baseFame + Math.Floor(bonus)) * 0.1;
 
             //Enemy of the Gods
-            if ((double) GodKills / (GodKills + MonsterKills) > 0.1)
+            if ((double)GodKills / (GodKills + MonsterKills) > 0.1)
                 bonus = Math.Floor(bonus) + (baseFame + Math.Floor(bonus)) * 0.1;
 
             //Slayer of the Gods
-            if ((double) GodKills / (GodKills + MonsterKills) > 0.5)
+            if ((double)GodKills / (GodKills + MonsterKills) > 0.5)
                 bonus = Math.Floor(bonus) + (baseFame + Math.Floor(bonus)) * 0.1;
 
             //Oryx Slayer
@@ -382,15 +382,15 @@ namespace LoESoft.Core
                 bonus = Math.Floor(bonus) + (baseFame + Math.Floor(bonus)) * 0.1;
 
             //Accurate
-            if ((double) ShotsThatDamage / Shots > 0.25)
+            if ((double)ShotsThatDamage / Shots > 0.25)
                 bonus = Math.Floor(bonus) + (baseFame + Math.Floor(bonus)) * 0.1;
 
             //Sharpshooter
-            if ((double) ShotsThatDamage / Shots > 0.5)
+            if ((double)ShotsThatDamage / Shots > 0.5)
                 bonus = Math.Floor(bonus) + (baseFame + Math.Floor(bonus)) * 0.1;
 
             //Sniper
-            if ((double) ShotsThatDamage / Shots > 0.75)
+            if ((double)ShotsThatDamage / Shots > 0.75)
                 bonus = Math.Floor(bonus) + (baseFame + Math.Floor(bonus)) * 0.1;
 
             //Explorer
@@ -425,7 +425,7 @@ namespace LoESoft.Core
                 if (chr.Items[i] == -1)
                     continue;
 
-                var b = data.Items[(ushort) chr.Items[i]].FameBonus;
+                var b = data.Items[(ushort)chr.Items[i]].FameBonus;
 
                 if (b > 0)
                     eq += (baseFame + Math.Floor(bonus)) * b / 100;
@@ -441,7 +441,7 @@ namespace LoESoft.Core
             else
                 firstBorn = false;
 
-            return (int) (baseFame + Math.Floor(bonus));
+            return baseFame + Math.Floor(bonus);
         }
     }
 }
